misc: Extract coupon logic from calculate fees service
